requirements.txt
openai-whisper>=20250625
sounddevice>=0.4.6
numpy>=1.24.0
pynput>=1.7.6
python-dotenv>=1.0.0

